requirements.txt
-r ./requirements/production.txt

django-heroku